Move Json => Base64 to Helper class
